Inform user about state of wallet-rpc download and monero wallet sync status

+++ src/main/cli/cli.ts
@@ -17,6 +17,11 @@ const queue = new PQueue({ concurrency: 1 });
 let cli: ChildProcessWithoutNullStreams | null = null;
 
 async function attemptKillMoneroWalletRpcProcess() {
+  if (process.env.SKIP_MONERO_WALLET_RPC_KILL === 'true') {
+    logger.debug('Skipping monero-wallet-rpc kill');
+    return;
+  }
+
   const WIN_COMMAND = `powershell.exe "Get-Process | Where-Object {$_.Path -like '*monero-wallet-rpc*'} | Stop-Process -Force"`;
   const UNIX_COMMAND = `ps aux | grep 'monero-wallet-rpc' | grep -v 'grep' | awk '{print $2}' | xargs kill -9`;
 
@@ -111,9 +116,7 @@ export async function spawnSubcommand(
                 );
               }
 
-              if (process.env.SKIP_MONERO_WALLET_RPC_KILL !== 'true') {
-                attemptKillMoneroWalletRpcProcess();
-              }
+              attemptKillMoneroWalletRpcProcess();
 
               cli = spawnProc(`./${binary.fileName}`, spawnArgs, {
                 cwd: binary.dirPath,
